bitboard.h: check for invalid square at isAttackedBy()
This commit Fixes the following crash scenario:

In Menus:
Click Edit | Setup Position...
Click Clear
Click on White King
CRASH

This commit checks in advance for InvalidSquare in function
is AttackedBy().

The crash was: If a King is removed its square is set to InvalidSquare,
Later when checking if that square is attacked we have a crash of accessing
index beyond the board size.